requirements/base.txt
-e git+https://github.com/metrasynth/radiant-voices@master#egg=radiant-voices
radiant-voices~=1.0.3
pytz~=2021.1  # https://github.com/stub42/pytz
python-slugify~=4.0.1  # https://github.com/un33k/python-slugify
Pillow~=8.1.2  # https://github.com/python-pillow/Pillow
rcssmin~=1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi~=20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise~=5.2.0  # https://github.com/evansd/whitenoise
redis~=3.5.3  # https://github.com/andymccurdy/redis-py
hiredis~=1.1.0  # https://github.com/redis/hiredis-py
celery~=4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat~=2.2.0  # https://github.com/celery/django-celery-beat
uvicorn[standard]~=0.13.4  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django~=3.1.7  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ~=0.4.5  # https://github.com/joke2k/django-environ
django-filter~=2.4.0
django-markdownify~=0.9.0
django-model-utils~=4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth~=0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms~=1.11.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor~=2.4  # https://github.com/django-compressor/django-compressor
django-redis~=4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework~=3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers~=3.7.0 # https://github.com/adamchainz/django-cors-headers
graphene~=2.1.8
graphene-django~=2.15.0  # https://docs.graphene-python.org/projects/django/en/latest/installation/

django-debug-toolbar~=3.2
requests~=2.25.1
pytest~=6.2.2
django-taggit~=1.3.0
django-vote~=2.2.0
